@@ -2,6 +2,6 @@
 
 #include "EventConnection.hpp"
 #include "EventSource.hpp"
-#include "GameEvents.hpp"
+#include "Game.hpp"
 #include "Keyboard.hpp"
 #include "Mouse.hpp"
