@@ -2,6 +2,12 @@
 
 #include "Running.hpp"
 #include "MainMenu.hpp"
-#include "GameState.hpp"
 #include "Statemachine.hpp"
 #include "GameOver.hpp"
+#include "GamePauze.hpp"
+#include "Cutscene.hpp"
+#include "SettingsMenu.hpp"
+
+/** @brief There must be a blank line at the end-of-file. Else the compilation will throw an 'unexpected EOF' error. */
+
+
