@@ -15,9 +15,20 @@
  */
 
 class Mouse : public SFMLEventDecoder {
-	sf::Window& window;
+	sf::Window* window;
 
 public:
+
+	/**
+	 * @fn	Mouse::Mouse()
+	 *
+	 * @brief	Default constructor.
+	 *
+	 * @author	Julian
+	 * @date	2018-01-19
+	 */
+
+	Mouse() : window(nullptr) { }
 
 	/**
 	 * @fn	Mouse::Mouse(sf::Window& window)
@@ -30,7 +41,7 @@ public:
 	 * @param [in,out]	window	The window to get the mouse posititions from.
 	 */
 
-	Mouse(sf::Window& window) : window(window) { }
+	Mouse(sf::Window& window) : window(&window) { }
 
 
 	/** @brief	Mouse left button up, fires when the left mouse button is released. */
@@ -47,6 +58,8 @@ public:
 
 	/** @brief	Mouse moved, fires when the mouse is moved. */
 	EventSource<sf::Vector2i> mouseMoved;
+
+	bool mouseLeftPressed = false, mouseRightPressed = false;
 
 	/**
 	 * @fn	virtual void SFMLEventDecoder::decodeSFMLEvent(sf::Event ev) const = 0;
@@ -66,19 +79,38 @@ public:
 	
 	void decodeSFMLEvent(sf::Event ev) const override {
 		if (ev.type == sf::Event::MouseMoved) {
-			mouseMoved.fire(sf::Mouse::getPosition(window));
+			mouseMoved.fire(sf::Mouse::getPosition(*window));
 		} else if (ev.type == sf::Event::MouseButtonPressed) {
 			if (ev.mouseButton.button == sf::Mouse::Left) {
-				mouseLeftButtonDown.fire(sf::Mouse::getPosition(window));
+				mouseLeftButtonDown.fire(sf::Mouse::getPosition(*window));
 			} else if (ev.mouseButton.button == sf::Mouse::Right) {
-				mouseRightButtonDown.fire(sf::Mouse::getPosition(window));
+				mouseRightButtonDown.fire(sf::Mouse::getPosition(*window));
 			}
 		} else if (ev.type == sf::Event::MouseButtonReleased) {
 			 if (ev.mouseButton.button == sf::Mouse::Left) {
-				 mouseLeftButtonUp.fire(sf::Mouse::getPosition(window));
+				 mouseLeftButtonUp.fire(sf::Mouse::getPosition(*window));
 			 } else if (ev.mouseButton.button == sf::Mouse::Right) {
-				 mouseRightButtonUp.fire(sf::Mouse::getPosition(window));
+				 mouseRightButtonUp.fire(sf::Mouse::getPosition(*window));
 			 }
 		}
+	}
+
+	/**
+	 * @fn	Mouse& Mouse::operator= (Mouse& rhs)
+	 *
+	 * @brief	Assignment operator
+	 *
+	 * @author	Julian
+	 * @date	2018-01-19
+	 *
+	 * @param [in,out]	rhs	The right hand side.
+	 *
+	 * @return	A shallow copy of this object.
+	 */
+
+	Mouse& operator= (const Mouse& rhs) {
+		window = rhs.window;
+
+		return *this;
 	}
 };
