@@ -1,10 +1,12 @@
 #pragma once
 
 #include <SFML/Graphics.hpp>
-#include "Drawable.hpp"
-#include "Collision.hpp"
 
-class PhysicsObject : public Drawable {
+#include "Drawable.hpp"
+#include "CollisionDetection.hpp"
+#include "Events.hpp"
+
+class PhysicsObject : public Drawable, public Collidable {
 protected:
 	sf::Transformable& transformable;
 	sf::Vector2f velocity = sf::Vector2f(0, 0);
@@ -13,7 +15,15 @@ protected:
 	PhysicsObject(sf::Shape& shape) : Drawable(shape), transformable(shape) { }
 	PhysicsObject(sf::Sprite& sprite) : Drawable(sprite), transformable(sprite) { }
 
+	bool sideCollision;
+
+	void onCollide(Collidable& other) override {
+		collided.fire(other);
+	}
+
 public:
+	EventSource<Collidable&> collided;
+
 	virtual void update(const float elapsedTime) {
 		applyForce(gravity * elapsedTime);
 		setPosition(getPosition() + velocity * elapsedTime);
@@ -43,70 +53,7 @@ public:
 		this->velocity = velocity;
 	}
 
-	virtual sf::Vector2f getVelocity() {
+	virtual sf::Vector2f getVelocity() const {
 		return velocity;
 	}
-
-	/**
-	* @brief Resolve an collision.
-	* @param other Rectangle.
-	*/
-	void resolveCollision(PhysicsObject &other) {
-		Collision collision = getCollision(other);
-
-		if (collision.check()) {
-			sf::Vector2f mvt;
-
-			if (collision.getIntersect().x > collision.getIntersect().y) {
-				if (collision.getDelta().x > 0.0f) {
-					mvt = sf::Vector2f(collision.getIntersect().x * 1.0f, 0.0f);
-				}
-				else {
-					mvt = sf::Vector2f(-collision.getIntersect().x * 1.0f, 0.0f);
-				}
-			}
-			else {
-				if (collision.getDelta().y > 0.0f) {
-					mvt = sf::Vector2f(0.0f, collision.getIntersect().y);
-				}
-				else {
-					mvt = sf::Vector2f(0.0f, -collision.getIntersect().y);
-					setVelocity({ getVelocity().x, 0 });
-				}
-				if (getVelocity().y > 0) {
-					setVelocity({ getVelocity().x, 0 });
-				}
-			}
-
-			setPosition(getPosition() + mvt);
-		}
-	}
-
-	/**
-	* @brief Get an collision instance containing the delta and intersect position between two rectangles.
-	* @param other Rectangle.
-	* @return Collision class instance.
-	*/
-	Collision getCollision(PhysicsObject &other) {
-		sf::Vector2f otherPosition = other.getPosition();
-		sf::Vector2f otherHalfSize = other.getSize() / 2.0f;
-		sf::Vector2f thisPosition = getPosition();
-		sf::Vector2f thisHalfSize = getSize() / 2.0f;
-
-		sf::Vector2f delta(otherPosition.x - thisPosition.x, otherPosition.y - thisPosition.y);
-		sf::Vector2f intersect(abs(delta.x) - (otherHalfSize.x + thisHalfSize.x), abs(delta.y) - (otherHalfSize.y + thisHalfSize.y));
-
-		return Collision(delta, intersect);
-	}
-
-	/**
-	* @brief Check if two rectangles intersects with each other.
-	* @param other Rectangle.
-	* @return Result.
-	*/
-	virtual bool intersects(PhysicsObject &other) {
-		Collision collision = getCollision(other);
-		return collision.check();
-	}
-
 };
