@@ -1,32 +1,54 @@
 #include <SFML/Graphics.hpp>
+#include <SFML/Audio.hpp>
 #include <iostream>
 
 #include "GameStates.hpp"
-#include "Button.hpp"
+#include "Events.hpp"
+#include "AssetManager.hpp"
+#include "fimmyIcon.hpp"
+
+//#define ENABLE_DEBUG_MODE
 
 int main() {
+	// Game constants.
 	const float FPS = 60.0f;
+	const bool  ENABLE_CUTSCENE = true;
 
-	sf::RenderWindow window(sf::VideoMode(1280, 720, 32), "Hello");
+	sf::RenderWindow window(sf::VideoMode(1280, 720, 32), "Fimmy the Game");
+
+	window.setIcon(sfml_icon.width, sfml_icon.height, sfml_icon.pixel_data);
+
+	window.setFramerateLimit(static_cast<unsigned int>(FPS));
+
+	game = Game(window);
+
+	// Load assets.
+	AssetManager::instance()->load("arial", "arial.ttf");
+
+	// Create an new statemachine.
 	Statemachine statemachine(window);
 
-	Button::setDefaultFont("arial.ttf");
+	// Set the default font.
+	Label::setDefaultFont(AssetManager::instance()->getFont("arial"));
 
-	// State definitions
+	// State definitions.
+	statemachine.registerState<GameOver>("game-over");
+	statemachine.registerState<Running>("running");
+	statemachine.registerState<MainMenu>("main-menu");
+	statemachine.registerState<GamePauze>("game-pauze");
+	statemachine.registerState<SettingsMenu>("settings-menu");
+	statemachine.registerState<Cutscene>("cutscene");
 
-	GameOver gameOver(statemachine);
-	Running running(statemachine);
-	MainMenu mainMenu(statemachine);
+	if (ENABLE_CUTSCENE) {
+		statemachine.doTransition("cutscene");
+	} else {
+		statemachine.doTransition("main-menu");
+	}
 
-
-	statemachine.doTransition("main-menu"); // initial state
-
-	// End state definitions
-
-	
-	window.setFramerateLimit((int) FPS);
+	// Clock used for frame timings.
 	sf::Clock clock;
 
+	// SFML event.
 	sf::Event ev;
 
 	while (window.isOpen()) {
@@ -37,12 +59,13 @@ int main() {
 				window.close();
 			}
 
-			statemachine.mouse.decodeSFMLEvent(ev);
-			statemachine.keyboard.decodeSFMLEvent(ev);
+			game.decodeSFMLEvent(ev);
 		}
 
 		if (elapsedTime >= 1.0f / FPS) {
+#ifdef ENABLE_DEBUG_MODE
 			window.setTitle(std::to_string(1 / elapsedTime));
+#endif // ENABLE_DEBUG_MODE
 			clock.restart();
 			statemachine.update(elapsedTime);
 		}
